Extract useState and useEffect for if edit component has features into hook
